+++ pkg/common/fileutils/fileutils.go
@@ -1,4 +1,4 @@
-package commonutils
+package fileutils
 
 import (
 	"fmt"
